requirements.txt
flask
pandas
numpy
matplotlib
scipy
openpyxl
PyYAML